initial multithread impl
